finish storage restAPI

+++ storage/localfs.py
@@ -11,27 +11,27 @@ from yuelibs import constants
 from yuelibs import utils
 
 class LocalFsStorage(object):
-    def __init__(self, path=''):
+    def __init__(self, path='', flag=False):
         self.path = path
-        self.delete()
-        self._create()
         self.meta = os.path.join(path, 'meta')
         self.image = os.path.join(path, 'image')
         self.template = os.path.join(path, 'template')
-        self.mount = self._getMount
+        if flag:
+            self.delete()
+            self._create()
+        self.mount = self._getMount()
         self.disk = self._getDevice()
         self.all,self.free = self._getStatus()
 
     def _create(self):
-        if self.path != None and os.path.exists(self.path):
+        if len(self.path)>0 and os.path.exists(self.path):
             pass
         else:
-            os.mkdir(self.path, mode=0644)
-
+            os.mkdir(self.path)
         if not os.path.exists(self.image):
-            os.mkdir(self.image, mode=0644)
+            os.mkdir(self.image)
         if not os.path.exists(self.template):
-            os.mkdir(self.template, mode=0644)
+            os.mkdir(self.template)
 
     def delete(self):
         try:
@@ -56,30 +56,36 @@ class LocalFsStorage(object):
     def _getStatus(self):
         cmd = 'df -h'
         out,err,errcode = utils.execShellCommand(cmd)
+        out = out.split('\n')
         for line in out:
-            list1 = line.split()
+            list1 = line.split(' ')
+            length = len(list1)
+            for i in range(1,length):
+                for elem in list1:
+                    if elem==' ' or len(elem)==0:
+                        list1.remove(elem)
             for elem in list1:
-                if elem == ' ':
-                    list1.remove(elem)
-
-            print list1
-
-            if list1[-1] == self.mount:
-                return list1[1],list1[3]
+                if list1[-1] == self.mount:
+                    return (list1[1],list1[3])
+        return (0,0)
 
     def _getDevice(self):
         cmd = 'lsblk -l'
         out,err,errcode = utils.execShellCommand(cmd)
+        out = out.split('\n')
+        disk = ''
         for line in out:
-            list1 = line.split()
+            list1 = line.split(' ')
             for elem in list1:
                 if elem == ' ':
                     list1.remove(elem)
 
-            print list1
+            if list1[-2] == 'disk':
+                disk = list1[0]
+                continue
 
             if list1[-1] == self.mount:
-                return list1[0]
+                return disk
 
 
     def _isMount(self, path):
@@ -87,12 +93,16 @@ class LocalFsStorage(object):
 
     def _getMount(self):
         path = self.path
-        for i in range(1, len(self.path)):
-            while not path.endswith("/"):
+        if path.endswith('/'):
+            path = self.path[:-1]
+        length = len(path)
+        for i in range(1, length):
+            if path.endswith("/"):
                 if len(path) > 1:
                     path = path[:-1]
+                    if len(path) == 1:
+                        return path
                     continue
-                else:
-                    break
-            if self.isMount(path):
+            if self._isMount(path):
                 return path
+        return '/'
